requirements.txt
Flask==2.3.2
Werkzeug==2.3.2
torch==2.1.0
torchvision==0.15.0
Pillow==9.4.0
